Mass restructure of fat module, renaming of cluster and read submod elements
Similar renaming to do throughout the rest of the module
